+++ lib/ex_daytona/connection.ex
@@ -118,7 +118,7 @@ defmodule ExDaytona.Connection do
         Keyword.get(
           tesla_options,
           :user_agent,
-          "openapi-generator - ExDaytona 0.2.0 - elixir"
+          "openapi-generator - ExDaytona 0.3.0 - elixir"
         )
       )